filters cars model

+++ src/entities/newcar/service/newcar.service.ts
@@ -73,11 +73,16 @@ export class NewCarService extends CrudService<typeof x> {
 
     async carModelVerification(car){
 
-        if(car.vin === '' || car.vin === null || car.vin.length !== 17){           return [{error: 'error en vin, no cumple con las condiciones == no nulo, no vacio, vin incompleto (17 caracteres) =='}, {car}]}
+        let carID = '';
+
+        if(car.vin) { carID = car.vin}
+        if(car.ID) { carID = car.ID }
+
+        if( carID === '' || carID === null || carID.length !== 17){                return [{error: 'error en identificador unico ( vin o ID), no cumple con las condiciones == no nulo, no vacio, vin 0 ID incompleto (17 caracteres) =='}, {car}]}
         if(car.agencyID === '' || car.agencyID === null){                          return [{error: 'sin agencyID'}, {car}]}
         if(car.brand === '' || car.brand === null){                                return [{error: 'sin brand'}, {car}]}
         if(car.model === '' || car.model === null || car.model.includes('/')){     return [{error: 'error en model, revise el modelo, no debe contener signo o caracteres especiales'}, {car}]}
-        if(car.series === '' || car.series === null || car.series.includes('/')){  return [{error: 'serie vacia o con caracteres especiales'}, {car}]}
+        if(car.series === '' || car.series === null ){                             return [{error: 'serie vacia'}, {car}]}
         if(car.price === '' || car.price === null) {                               return [{error: 'sin precio'}, {car}]}
         if(car.chassisType === '' || car.chassisType === null){                    return [{error: 'sin segmento'}, {car}]}
         if(car.year === '' || car.year === null){                                  return [{error: 'sin año, verifique los datos ingresados'}, {car}]}
@@ -231,6 +236,7 @@ export class NewCarService extends CrudService<typeof x> {
             let carlist = await this.repository.findAll();
 
             let carinlist = [];
+            let carlistban = []
 
             for (let response of responses) {
                 if (response.data.success) {
@@ -251,8 +257,10 @@ export class NewCarService extends CrudService<typeof x> {
 
                         })
 
+                       let verificacion = await this.carModelVerification(sc)
 
-                        if (sc.isAvailable === 'S' && sc.isReserved === 'N' && sc.demo !== 'S') {
+
+                        if (sc.isAvailable === 'S' && sc.isReserved === 'N' && sc.demo !== 'S' && verificacion === 200) {
 
                             let newmodel: string;
                             let MetaDescription: string;
@@ -328,27 +336,31 @@ export class NewCarService extends CrudService<typeof x> {
                             }
                             if (BDID !== '') {
 
-                                await this.repository.update(BDID, newCar)
+                                // await this.repository.update(BDID, newCar)
                                 updateitem++
                             } else {
-                                newCarsArray.push(newCar)
+                                // newCarsArray.push(newCar)
                             }
 
 
                         }
                         else {
 
-                            carlist.items.forEach((car: any) => {
+                            if(verificacion !== 200){
+                                carlistban.push(verificacion)
+                            }
+                            else {
+                                carlist.items.forEach((car: any) => {
 
-                                if (sc.ID === car.vin) {
-
-                                    this.finishedcar.create(car)
-
-                                    console.log( 'auto descartado: ', car.vin)
-                                    this.repository.delete(car._id)
-                                }
-                            })
-
+                                    if (sc.ID === car.vin) {
+    
+                                        this.finishedcar.create(car)
+    
+                                        console.log( 'auto descartado: ', car.vin)
+                                        this.repository.delete(car._id)
+                                    }
+                                })
+                            }
                         }
                     }
                 }
@@ -396,8 +408,10 @@ export class NewCarService extends CrudService<typeof x> {
             }
         
             return {
+                banCarlist: carlistban,
                 count: newCarsArray.length,
-                results: createdCars
+                results: createdCars,
+                
             }
         } catch (err) {
             console.log('error en update newcar: ' + err)
